fix(GameSceneView): Make AddAccessibilityDirectInteractionTrait custom modifier
