+++ WavePong/Views/AddAccessibilityDirectInteractionTrait.swift
@@ -1,0 +1,22 @@
+//
+//  AddAccessibilityDirectInteractionTrait.swift
+//  WavePong
+//
+//  Created by rsbj on 23/08/23.
+//
+
+import SwiftUI
+
+struct AddAccessibilityDirectInteractionTrait: ViewModifier {
+    var condition: Bool
+    
+    func body(content: Content) -> some View {
+        if condition { // assuming you have a condition here
+            return content
+                .accessibilityAddTraits(.allowsDirectInteraction)
+        } else {
+            return content
+                .accessibilityRemoveTraits(.allowsDirectInteraction)
+        }
+    }
+}
